[IMP] iwp: mail to structure for new loan received
