+++ investor_wallet_platform_base/data/mail_template_data.xml
@@ -385,6 +385,25 @@
             ]]></field>
         </record>
 
+        <record id="investor_wallet_platform_base.loan_subscription_received" model="mail.template">
+            <field name="name">Loan Subscription Request Received - Template</field>
+            <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+            <field name="subject">Loan subscription received (Ref ${object.loan_issue_id.name or 'n/a'})</field>
+            <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+            <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+            <field name="model_id" ref="easy_my_coop_loan.model_loan_issue_line"/>
+            <field name="auto_delete" eval="True"/>
+            <field name="lang">${object.partner_id.lang}</field>
+            <field name="iwp" eval="True"/>
+            <field name="body_html"><![CDATA[
+<div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
+
+    <p>A new loan subscription  was received: ${object.loan_issue_id.name} - ${object.partner_id.name}</p>
+
+</div>
+            ]]></field>
+        </record>
+
         <record id="investor_wallet_platform_base.loan_subscription_confirmation" model="mail.template">
             <field name="name">Loan Subscription Confirmation Email - Template</field>
             <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
