+++ investor_wallet_platform_base/demo/mail_template_demo.xml
@@ -14,7 +14,7 @@
     <record id="res_partner_structure_coopiteasy_demo" model="res.partner">
         <field name="mail_serveur_out" ref="mail_server_out_coopiteasy_demo"/>
     </record>
-    
+
     <record id="email_template_confirmation_coopiteasy_demo" model="mail.template">
         <field name="name">Confirmation Email</field>
         <field name="mail_server_id" ref="mail_server_out_coopiteasy_demo"/>
@@ -193,7 +193,7 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Loan Issue Payment Request</p>
-        ]]></field>    
+        ]]></field>
     </record>
 
 
@@ -242,7 +242,7 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Company Confirmation</p>
-        ]]></field> 
+        ]]></field>
     </record>
 
     <record id="email_template_release_capital_coopcity_demo" model="mail.template">
@@ -284,7 +284,7 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Payment Received Confirmation</p>
-        ]]></field> 
+        ]]></field>
     </record>
 
     <record id="email_template_certificat_increase_coopcity_demo" model="mail.template">
@@ -305,7 +305,7 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Share Increase</p>
-        ]]></field>  
+        ]]></field>
     </record>
 
     <record id="email_template_share_transfer_coopcity_demo" model="mail.template">
@@ -326,7 +326,7 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Share transfer</p>
-        ]]></field>   
+        ]]></field>
     </record>
 
     <record id="email_template_share_update_coopcity_demo" model="mail.template">
@@ -347,7 +347,7 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Share update</p>
-        ]]></field>   
+        ]]></field>
     </record>
 
     <record id="loan_subscription_confirmation_coopcity_demo" model="mail.template">
@@ -366,7 +366,7 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Loan Subscription Confirmation Email</p>
-        ]]></field>    
+        ]]></field>
     </record>
 
     <record id="loan_issue_payment_request_coopcity_demo" model="mail.template">
@@ -385,7 +385,51 @@
         <field name="body_html"><![CDATA[
         <div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
             <p>Hello ${object.name}, Loan Issue Payment Request</p>
-        ]]></field>    
+        ]]></field>
+    </record>
+
+    <record id="mail_loan_subscription_received_coopiteasy_demo" model="mail.template">
+        <field name="name">Loan Subscription Request Received</field>
+        <field name="mail_server_id" ref="mail_server_out_coopiteasy_demo"/>
+        <field name="structure" ref="res_partner_structure_coopiteasy_demo"/>
+        <field name="template_key">loan_subscription_received</field>
+        <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+        <field name="subject">Loan subscription received (Ref ${object.loan_issue_id.name or 'n/a'})</field>
+        <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+        <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+        <field name="model_id" ref="easy_my_coop_loan.model_loan_issue_line"/>
+        <field name="auto_delete" eval="True"/>
+        <field name="lang">${object.partner_id.lang}</field>
+        <field name="iwp" eval="True"/>
+        <field name="body_html"><![CDATA[
+<div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
+
+<p>A new loan subscription  was received: ${object.loan_issue_id.name} - ${object.partner_id.name}</p>
+
+</div>
+        ]]></field>
+    </record>
+
+    <record id="mail_loan_subscription_received_coopcity_demo" model="mail.template">
+        <field name="name">Loan Subscription Request Received</field>
+        <field name="mail_server_id" ref="mail_server_out_coopcity_demo"/>
+        <field name="structure" ref="res_partner_structure_coopcity_demo"/>
+        <field name="template_key">loan_subscription_received</field>
+        <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+        <field name="subject">Loan subscription received (Ref ${object.loan_issue_id.name or 'n/a'})</field>
+        <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+        <field name="email_from">${(object.structure.email or object.loan_issue_id.user_id.email)|safe}</field>
+        <field name="model_id" ref="easy_my_coop_loan.model_loan_issue_line"/>
+        <field name="auto_delete" eval="True"/>
+        <field name="lang">${object.partner_id.lang}</field>
+        <field name="iwp" eval="True"/>
+        <field name="body_html"><![CDATA[
+<div style="font-family: 'Lucica Grande', Ubuntu, Arial, Verdana, sans-serif; font-size: 12px; color: rgb(34, 34, 34); background-color: #FFF; ">
+
+<p>A new loan subscription  was received: ${object.loan_issue_id.name} - ${object.partner_id.name}</p>
+
+</div>
+        ]]></field>
     </record>
 
 </odoo>
