+++ investor_wallet_platform_base/models/loan_issue.py
@@ -85,43 +85,3 @@ class LoanIssue(models.Model):
     def refuse(self):
         for issue in self:
             issue.validation_requested = False
-
-
-class LoanIssueLine(models.Model):
-    _inherit = "loan.issue.line"
-
-    structure = fields.Many2one(
-        comodel_name="res.partner",
-        string="Platform Structure",
-        related="loan_issue_id.structure",
-        store=True,
-    )
-    creation_mode = fields.Selection(
-        [("auto", "Automatic"), ("manual", "Manual")],
-        string="Creation mode",
-        default="auto",
-        readonly=True,
-        help=(
-            "'auto' means created by system, 'manual' means manually "
-            "created by a user."
-        ),
-    )
-
-    def get_loan_sub_mail_template(self):
-        template_obj = self.env["mail.template"]
-        return template_obj.get_email_template_by_key(
-            "loan_sub_conf", self.structure
-        )
-
-    def get_loan_pay_req_mail_template(self):
-        template_obj = self.env["mail.template"]
-        return template_obj.get_email_template_by_key(
-            "loan_payment_req", self.structure
-        )
-
-    @api.multi
-    def get_confirm_paid_email_template(self):
-        self.ensure_one()
-        return self.env["mail.template"].get_email_template_by_key(
-            "loan_payment_received", self.structure
-        )
